Implement support for TIMING-MARK option.
Attempt to use the TELNET protocol TIMING-MARK option (RFC 860) to
implement end-to-end acknowledgements, if possible.  Supporting TELNET
implementations must respond to IAC DO TIMING-MARK with an appropriate
response of IAC WILL TIMING-MARK or IAC WON'T TIMING-MARK, depending on
whether or not the implementation is providing assurance that all output
before the IAC DO TIMING-MARK has been "completely processed, printed,
discarded, or otherwise disposed of."

Unlike normal TELNET options, this particular option is stateless, and
supporting implementations are required to respond to EACH request with
an appropriate response.  Furthermore, conforming TELNET implementations
which do NOT understand this option are ALSO required to respond to each
request (with IAC WON'T TIMING-MARK), since this is a requirement of the
base TELNET protocol for unrecognized options.  As such, any conforming
TELNET client should respond, whether or not it understands the option.

Of course, in the real world, a client may not correctly implement the
TELNET protocol, or a user might be connecting with a simple TCP client
that doesn't support the TELNET protocol at all.  Therefore, attempt to
use the IAC DO TIMING-MARK sequence and remember whether or not the
client responded to it, before relying on this mechanism for end-to-end
acknowledgements.

If this mechanism appears to be unusable, disable the use of this option
and simulate an "acknowledgement" immediately after the preceding output
has been fully buffered by the kernel.  This doesn't provide end-to-end
confirmation that the output has been received by the client, but it
will have to do for any clients which don't respond to the option.