Echo handling improvements and raw TCP support.
Add "probably" to change warning to "password probably WILL echo", to
reflect the possibility that it actually won't.  Modify banner to send
the "Welcome to Phoenix!" part immediately, while still waiting for
option negotiation to finish before sending the rest.

Add support for raw TCP connections -- if a line of input is received
before option negotiation is finished, assume raw TCP and continue with
login.  This can fail with telnet if a line is typed before telnet opens
the connection, but that's hard to avoid.

Add /set echo command primarily for raw TCP users; this is somewhat of a
kludge -- normally /set is processed by the Session class.  I'm not sure
of a cleaner way to do this without making something ugly like
%commands, so I guess this will have to do for now.