Add "most.cc" file to simplify compilation process.
Modify to build with "most.cc" module, which simply includes all modules
that include "phoenix.h" -- compiler spends most time on header files,
not module code, so this is faster and help eliminate highly-redundant
debug information in the final executable.  Also add preprocessor
directives to "phoenix.h" header, so it can be included multiple times.
So much for separate compilation...