Add Object reference count kludge.
Modify DeleteReference() to set a negative reference count to allow for
some temporary pointers to the object being destructed, without calling
the destructors again in an endless loop when such temporary pointers go
out of scope.  Modify the destructor sanity check to only check for a
positive reference count, due to the above modification.