Update Object/Pointer<Type> smart pointer classes.
Refactored original 1994 code to work correctly with GCC 4.8.3, which
caused the original code to crash when it tried to "delete this" from
Object::DeleteReference() on objects with virtual methods, because
"this" had a different address in the Object class than "this" in the
derived class had, for the same object.

This updated implementation adds Pointer::SetPointer() for consistent,
simplified and more reliable pointer assignments.  Unreferenced objects
are now deleted by this new method instead of Object::DeleteReference(),
and assignments of objects being destroyed are now automatically set to
NULL instead to prevent destructor loops.