Ignore errors returned from respond_to.send().
Since respond_to is a oneshot channel, it should only fail if the other
end of the channel has already been closed.  If that happens, send()
will return the value that was passed in, not an error.  There is no
reasonable recovery strategy here for a failure to respond to a request
because the requestor is gone.  Logging a debugging message here would
be an option, but it seems like overkill for an unlikely race condition
that probably isn't even interesting to see in the logs even if it does
happen to occur.