Fix handle pattern for Name type.
Name objects are intended to be immutable, so there was no need to
include RwLock in the structure.  Remove RwLock and keep only Arc (since
names do need to be shared across threads), fix the constructors to
create the handle wrapper object properly, and fix all field references
to use "self.0" instead of "self" to access the fields.